WhoComesFirst
Player A and Player B will compete to reach a destination. The first player to reach the target point wins the game. 0, starting point; 100 is the destination (the winning player can exceed 100 on the last roll of the dice). By rolling a dice, players advance according to the number of dice. The algorithm works according to the following assumptions. When players come to a two-digit number, they have to go back 1 step if the number is prime. When it comes to a number that is divisible by 10, players move forward 2 steps. The program shows how many times the dice were rolled when the game was completed and how many steps the winning player finished the game.